+++ lib/rubocop.rb
@@ -119,6 +119,7 @@ require 'rubocop/cop/style/when_then'
 require 'rubocop/cop/style/while_until_do'
 require 'rubocop/cop/style/word_array'
 
+require 'rubocop/cop/rails/has_and_belongs_to_many'
 require 'rubocop/cop/rails/validation'
 
 require 'rubocop/formatter/base_formatter'